+++ changelog.md
@@ -8,6 +8,7 @@
 - \[mpm\] Add new command to export installed packages to a SBOM file in SPDX or CycloneDX standard. Closes {issue}`936`.
 - \[mpm\] Add new dependencies on `spdx-tools` and `cyclonedex-python-lib`.
 - \[mpm\] Update list of recognized pURL scheme types.
+- \[apt\] Add architecture in package metadata.
 
 ## [5.17.0 (2024-07-07)](https://github.com/kdeldycke/meta-package-manager/compare/v5.16.0...v5.17.0)